requirements.txt
absl-py==1.2.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
bidict==0.22.0
cachetools==5.2.0
certifi
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.1
flatbuffers==2.0
fonttools==4.34.4
gast==0.5.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
librosa==0.9.1
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numba==0.56.0
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.2.0
pooch==1.6.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///E:/python_program/SR/speech/packages/PyAudio-0.2.12-cp38-cp38-win_amd64.whl
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
python-speech-features==0.6
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.0
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
gunicorn==20.1.0